+++ src/main/java/de/ellpeck/actuallyadditions/mod/config/CommonConfig.java
@@ -89,13 +89,11 @@ public class CommonConfig {
     }
 
     public static class Other {
-        public static ModConfigSpec.BooleanValue HIDE_ENERGY_OVERLAY;
         public static ModConfigSpec.BooleanValue SOLID_XP_ALWAYS_ORBS;
         public static ModConfigSpec.BooleanValue DO_UPDATE_CHECK;
         public static ModConfigSpec.BooleanValue UPDATE_CHECK_VERSION_SPECIFIC;
         public static ModConfigSpec.BooleanValue DO_CAT_DROPS;
         public static ModConfigSpec.BooleanValue DO_BAT_DROPS;
-        public static ModConfigSpec.IntValue FUR_CHANCE;
         public static ModConfigSpec.BooleanValue WORMS;
         public static ModConfigSpec.IntValue WORMS_DIE_TIME;
         public static ModConfigSpec.BooleanValue CTRL_EXTRA_INFO;
@@ -117,9 +115,6 @@ public class CommonConfig {
 
             BUILDER.comment("Everything else").push("other");
 
-            HIDE_ENERGY_OVERLAY = BUILDER.comment("If true, when looking at blocks the Energy Overlay will be hidden.")
-                .define("hideEnergyOverlay", false);
-
             SOLID_XP_ALWAYS_ORBS = BUILDER.comment("If true, Solidified Experience will always spawn orbs, even for regular players.")
                 .define("solidXPOrbs", false);
 
@@ -134,9 +129,6 @@ public class CommonConfig {
 
             DO_BAT_DROPS = BUILDER.comment("Should Bat wings drop from Bats?")
                 .define("doBatDrops", true);
-
-            FUR_CHANCE = BUILDER.comment("The 1/n drop chance, per tick, for a fur ball to be dropped.")
-                .defineInRange("furDropChance", 5000, 1, Integer.MAX_VALUE);
 
             WORMS = BUILDER.comment("If true, worms will drop from tilling the soil.")
                 .define("tillingWorms", true);